requirements.txt
apache-airflow-providers-postgres==4.1.0
sqlalchemy
pandas
psycopg2
pyspark